+++ src/units/constants.rs
@@ -6,4 +6,4 @@ const GRAVITY_CONSTANT_DIMENSION: Dimension = Dimension {
     time: -2,
     mass: -1,
 };
-pub const GRAVITY_CONSTANT: Quantity<f32, GRAVITY_CONSTANT_DIMENSION> = Quantity(6.67430e-11);
+pub const GRAVITY_CONSTANT: Quantity<f64, GRAVITY_CONSTANT_DIMENSION> = Quantity(6.67430e-11);
